+++ src/main/java/net/nwrn/pf_contest/exception/ErrorController.java
@@ -1,4 +1,4 @@
-package net.nwrn.pf_contest;
+package net.nwrn.pf_contest.exception;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -6,7 +6,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import lombok.RequiredArgsConstructor;
-import net.nwrn.pf_contest.exception.ExceptionService;
 
 @RequiredArgsConstructor
 @Controller
