+++ src/main/java/net/nwrn/pf_contest/security/AuthorizationService.java
@@ -1,13 +1,14 @@
-package net.nwrn.pf_contest;
+package net.nwrn.pf_contest.security;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import net.nwrn.pf_contest.users.repository.UserEntity;
+import net.nwrn.pf_contest.users.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.Base64;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
