+++ src/main/java/net/nwrn/pf_contest/users/repository/UserRepository.java
@@ -1,9 +1,8 @@
-package net.nwrn.pf_contest;
+package net.nwrn.pf_contest.users.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
