+++ src/main/java/net/nwrn/pf_contest/users/service/UserService.java
@@ -1,8 +1,10 @@
-package net.nwrn.pf_contest;
+package net.nwrn.pf_contest.users.service;
 
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface UserService {
     public void join(String userId, String password, HttpServletResponse response);
     public void login(String userId, String password, HttpServletResponse response);
+    public void logout(Long id, HttpServletResponse response);
+
 }
